Initial improvement.
Converted code based on HEX instead of BinaryArrays. Not sure how it will perform but it sure is easier to program and works on other platforms.

Code didn't work properly on any platform. Some fixes are made though some functionality is broken. I expect it to be fixed soon. A2S_INFO is working and so are some other base functions now.
Still needs replacing to new features